Rename VLA entrypoint modules
