+++ python/sglang/multimodal_gen/runtime/entrypoints/http_server.py
@@ -30,7 +30,8 @@ from sglang.multimodal_gen.runtime.entrypoints.utils import (
     prepare_request,
     save_outputs,
 )
-from sglang.multimodal_gen.runtime.entrypoints.vla import action_api, openpi
+from sglang.multimodal_gen.runtime.entrypoints.vla import api as vla_api
+from sglang.multimodal_gen.runtime.entrypoints.vla import openpi
 from sglang.multimodal_gen.runtime.scheduler_client import async_scheduler_client
 from sglang.multimodal_gen.runtime.server_args import ServerArgs, get_global_server_args
 from sglang.multimodal_gen.runtime.server_warmup import (
@@ -402,7 +403,7 @@ def create_app(server_args: ServerArgs):
     app.include_router(video_api.router)
     app.include_router(realtime_video_api.router)
     if server_args.pipeline_config.task_type.is_action_gen():
-        app.include_router(action_api.router)
+        app.include_router(vla_api.router)
         app.include_router(openpi.router)
     app.include_router(mesh_api.router)
     app.include_router(weights_api.router)
