+++ python/sglang/multimodal_gen/runtime/entrypoints/vla/protocol.py
